circ
Started as a function to quickly draw circles but is much more functional now. 

Specifying "r" adjusts the radius of the figure
Specifying "x" "y" or "cen" adjusts the x and y coordinates.
Specifying "n" determines the number of sides for the figure. Small values of "n" appear as polygons.
Specifying "arc" adjusts how much of the figure is drawn, value is given in radians.
Specifying "a" and "b" stretches the figure along the horizontal and vertical, producing an ellipse or a polygon inscribe within an ellipse.
Specifying "p" changes the phase of the drawing.
Specifying "rotate" rotates an ellipse or polygon around its center.