Update dice
A method for calculating the probability of a given sum when rolling multiple dice. The user must specify "n", the number of dice to be rolled, and either "s" the number of sides if the side are sequential integers and start at 1 or "faces" for other arrangements. For faces all sides must be integer valued.